Adjusting file naming conventions
