+++ src/lgfx/panel/Panel_ILI9163.hpp
@@ -1,7 +1,7 @@
 #ifndef LGFX_PANEL_ILI9163_HPP_
 #define LGFX_PANEL_ILI9163_HPP_
 
-#include "panel_ilitek_common.hpp"
+#include "PanelIlitekCommon.hpp"
 
 namespace lgfx
 {
