+++ src/lgfx/panel/Panel_ILI9486.hpp
@@ -1,22 +1,20 @@
-#ifndef LGFX_PANEL_ILI9488_HPP_
-#define LGFX_PANEL_ILI9488_HPP_
+#ifndef LGFX_PANEL_ILI9486_HPP_
+#define LGFX_PANEL_ILI9486_HPP_
 
-#include "panel_ilitek_common.hpp"
+#include "PanelIlitekCommon.hpp"
 
 namespace lgfx
 {
-  struct Panel_ILI9488 : public PanelIlitekCommon
+  struct Panel_ILI9486 : public PanelIlitekCommon
   {
-    Panel_ILI9488(void)
+    Panel_ILI9486(void)
     {
       panel_width  = memory_width  = 320;
       panel_height = memory_height = 480;
 
-      freq_write = 40000000;
-      freq_read  = 20000000;
-      freq_fill  = 40000000;
-      read_depth  = rgb888_3Byte;
-      write_depth = rgb888_3Byte;
+      freq_write = 27000000;
+      freq_read  = 16000000;
+      freq_fill  = 27000000;
     }
 
   protected:
@@ -28,48 +26,31 @@ namespace lgfx
       static constexpr std::uint8_t FRMCTR3 = 0xB3;
       static constexpr std::uint8_t INVCTR  = 0xB4;
       static constexpr std::uint8_t DFUNCTR = 0xB6;
-      static constexpr std::uint8_t ETMOD   = 0xB7;
       static constexpr std::uint8_t PWCTR1  = 0xC0;
       static constexpr std::uint8_t PWCTR2  = 0xC1;
       static constexpr std::uint8_t PWCTR3  = 0xC2;
       static constexpr std::uint8_t PWCTR4  = 0xC3;
       static constexpr std::uint8_t PWCTR5  = 0xC4;
-      static constexpr std::uint8_t VMCTR   = 0xC5;
+      static constexpr std::uint8_t VMCTR1  = 0xC5;
+      static constexpr std::uint8_t VMCTR2  = 0xC7;
       static constexpr std::uint8_t GMCTRP1 = 0xE0; // Positive Gamma Correction
       static constexpr std::uint8_t GMCTRN1 = 0xE1; // Negative Gamma Correction
-      static constexpr std::uint8_t ADJCTL3 = 0xF7;
     };
-
     const std::uint8_t* getInitCommands(std::uint8_t listno) const override {
       static constexpr std::uint8_t list0[] = {
+          CMD::SLPOUT,  CMD_INIT_DELAY, 5,    // Exit sleep mode
+          CMD::PWCTR3 , 1, 0x44,       // Power control 3
+          CMD::VMCTR1 , 4, 0x00, 0x00, 0x00, 0x00,
 
-          CMD::PWCTR1,  2, 0x17,  // VRH1
-                           0x15,  // VRH2
-          CMD::PWCTR2,  1, 0x41,  // VGH, VGL
-          CMD::VMCTR ,  3, 0x00,  // nVM
-                           0x12,  // VCM_REG
-                           0x80,  // VCM_REG_EN
-          CMD::COLMOD,  1, 0x66,  // Interface Pixel Format = 18bit
-          CMD::FRMCTR1, 1, 0xA0,  // Frame rate = 60Hz
-          CMD::INVCTR,  1, 0x02,  // Display Inversion Control = 2dot
-          CMD::DFUNCTR, 2, 0x02,  // Nomal scan
-                           0x02,  // 5 frames
-          CMD::ETMOD,   1, 0xC6,  // 
-          CMD::ADJCTL3, 4, 0xA9,  // Adjust Control 3 
-                           0x51,
-                           0x2C,
-                           0x82,
+          CMD::GMCTRP1,15, 0x0F, 0x1F, 0x1C, 0x0C, 0x0F, 0x08, 0x48, 0x98,
+                           0x37, 0x0A, 0x13, 0x04, 0x11, 0x0D, 0x00,
 
-//          CMD::GMCTRP1,15, 0x00, 0x03, 0x09, 0x08, 0x16, 0x0A, 0x3F, 0x78,
-//                           0x4C, 0x09, 0x0A, 0x08, 0x16, 0x1A, 0x0F, 
-//          CMD::GMCTRN1,15, 0x00, 0x16, 0x19, 0x03, 0x0F, 0x05, 0x32, 0x45,
-//                           0x46, 0x04, 0x0E, 0x0D, 0x35, 0x37, 0x0F, 
-
+          CMD::GMCTRN1,15, 0x0F, 0x32, 0x2E, 0x0B, 0x0D, 0x05, 0x47, 0x75,
+                           0x37, 0x06, 0x10, 0x03, 0x24, 0x20, 0x00,
           0xFF,0xFF, // end
       };
       static constexpr std::uint8_t list1[] = {
-          CMD::SLPOUT,  CMD_INIT_DELAY, 120,    // Exit sleep mode
-          CMD::DISPON,  CMD_INIT_DELAY,  20,    // Set display on
+          CMD::DISPON, 0,     // Set display on
           0xFF,0xFF, // end
       };
       switch (listno) {
@@ -92,8 +73,6 @@ namespace lgfx
       r = (((r + offset_rotation) & 3) | (r & 4));
       return madctl_table[r];
     }
-
-    std::uint8_t getColMod(std::uint8_t) const override { return RGB666_3BYTE; }
   };
 }
