+++ src/lgfx/panel/Panel_ST7789.hpp
@@ -1,7 +1,7 @@
 #ifndef LGFX_PANEL_ST7789_HPP_
 #define LGFX_PANEL_ST7789_HPP_
 
-#include "panel_ilitek_common.hpp"
+#include "PanelIlitekCommon.hpp"
 
 namespace lgfx
 {
